[ScrollLabel.py] fix "scrollbarSliderPicture" and "scrollbarBackgroundPicture" (#2820)
Previously these attributes did not work and raised an AttributeError.

Switch "self.skinAttributes" to "self.skinAttributes[:]" so we are iterating a copy of the list while modifying the original.

Previously the loop which was iterating "self.skinAttributes" was removing items from the same list (a Python no-no). This means that if the  current (or a previous) item is deleted from the sequence, the next item will be skipped (since it gets the index of the current item which has already been treated).