With left/right you can now zap through parental controll question
During zap when you get on a parental controll channel you get a popup.
You could not zap through it when using the left/right buttons. The only
way to zap over these channels was go via the channelselection box in
case you do not want to enter a pincode or go back to the previous
channel via exit. With this change you can zap to the next or previous
channel with the left/right button.

Consider 2DO... Add the buttons for neutrino mode to the InputPin class and
arrange that it correctly respond when we're in Neutrino mode. Maybo
also add channel up/down buttons when they configured as zap buttons.